+++ CHANGELOG.md
@@ -11,6 +11,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 #### Fixed
 - Fixed broken office selector on players page
 - Bot waiting on initial throw
+- Prevent spamming of 'order_changes' when changing order
 
 
 ## [2.3.0] - 2022-03-06